refactor(voice): Phase 1 — backend abstraction layer
- Define SttBackend and TtsBackend traits in backend/mod.rs
- Create KokoroBackend (TTS) in backend/kokoro.rs
- Create WhisperBackend (STT) in backend/whisper.rs
- Convert tts.rs and stt.rs to backward-compatible re-export shims
- Add feature flags: kokoro, whisper, coreml, cuda
- Refactor pipeline.rs to use Box<dyn SttBackend> / Box<dyn TtsBackend>
- Add async-trait dependency
- All 25 tests pass, downstream gglib-tauri compiles
