+++ crates/gglib-voice/Cargo.toml
@@ -24,7 +24,11 @@ missing_errors_doc = "allow"
 missing_panics_doc = "allow"
 
 [features]
-default = []
+default = ["kokoro", "whisper"]
+# Kokoro v1.0 TTS backend (via vendored kokoro-tts fork)
+kokoro = ["dep:kokoro-tts", "dep:ort"]
+# Whisper STT backend (via whisper-rs / whisper.cpp)
+whisper = ["dep:whisper-rs"]
 # Apple Silicon acceleration via CoreML for ONNX Runtime (Kokoro TTS)
 coreml = ["ort/coreml"]
 # CUDA acceleration
@@ -47,6 +51,9 @@ tracing = { workspace = true }
 # Serialization
 serde = { workspace = true }
 
+# Async traits (for TtsBackend dyn compatibility)
+async-trait = { workspace = true }
+
 # Audio capture (cross-platform: CoreAudio/WASAPI/ALSA)
 cpal = "0.15"
 
@@ -56,22 +63,22 @@ rodio = { version = "0.20", default-features = false, features = ["wav"] }
 # Audio resampling (device sample rate → 16kHz for whisper)
 rubato = "0.16"
 
-# Speech-to-Text: whisper.cpp Rust bindings
-whisper-rs = "0.15"
+# Speech-to-Text: whisper.cpp Rust bindings (optional — feature "whisper")
+whisper-rs = { version = "0.15", optional = true }
 
-# Text-to-Speech: Kokoro v1.0 — local fork with Misaki G2P patch.
+# Text-to-Speech: Kokoro v1.0 — local fork with Misaki G2P patch (optional — feature "kokoro").
 # The upstream kokoro-tts crate uses an espeak dictionary for G2P, but the
 # Kokoro model was trained with the Misaki G2P engine. The phoneme mismatch
 # causes garbled/hallucinated speech. Our fork patches the G2P layer to use
 # misaki-rs (the correct engine) while keeping all ONNX inference untouched.
-kokoro-tts = { path = "../kokoro-tts-custom", features = ["misaki-g2p"] }
+kokoro-tts = { path = "../kokoro-tts-custom", features = ["misaki-g2p"], optional = true }
 
-# ONNX Runtime (used by kokoro-tts, configurable acceleration)
+# ONNX Runtime (used by kokoro-tts, configurable acceleration — optional, pulled in by "kokoro" feature)
 ort = { version = "2.0.0-rc.11", default-features = false, features = [
     "download-binaries",
     "copy-dylibs",
     "ndarray",
-] }
+], optional = true }
 
 # HTTP client for model downloads
 reqwest = { workspace = true }
