+++ crates/gglib-voice/src/lib.rs
@@ -3,6 +3,7 @@
 #![deny(unused_crate_dependencies)]
 
 // ort is used transitively by kokoro-tts features (coreml, cuda)
+#[cfg(feature = "kokoro")]
 use ort as _;
 
 #[cfg(test)]
@@ -10,6 +11,7 @@ use tempfile as _;
 #[cfg(test)]
 use tokio_test as _;
 
+pub mod backend;
 pub mod capture;
 pub mod error;
 pub mod gate;
@@ -26,3 +28,6 @@ pub use error::VoiceError;
 pub use gate::EchoGate;
 pub use models::{SttModelInfo, TtsModelInfo, VoiceModelCatalog, VoiceModelId};
 pub use pipeline::{VoiceEvent, VoicePipeline, VoicePipelineConfig, VoiceState};
+
+// Re-export backend trait types at crate root for ergonomic imports
+pub use backend::{SttBackend, SttConfig, TtsAudio, TtsBackend, TtsConfig, VoiceGender, VoiceInfo};
